readcount, replycount, likecount를 int에서 long으로 변경

+++ realSemiProject/src/main/java/com/semi/project/dto/BoardDto.java
@@ -12,7 +12,7 @@ String memberId;
 String boardTitle;
 String boardContent;
 Date boardCtime;
-int boardReadcount;
-int boardReplycount;
-int boardLikecount;
+long boardReadcount;
+long boardReplycount;
+long boardLikecount;
 }
